// frontend/src/types.ts
export interface RankingItem {
  id: string
  name: string
  value: number
  rank: number
  extra?: any
}